error creating an extra fork

+++ src/tools2.c
@@ -6,7 +6,7 @@
 /*   By: marvin <marvin@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/13 14:24:21 by marvin            #+#    #+#             */
-/*   Updated: 2024/12/31 17:34:47 by marvin           ###   ########.fr       */
+/*   Updated: 2025/01/11 18:38:59 by marvin           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 
 void	print_thread(t_philo *philos, int id, char *str)
 {
-	int	time;
+	size_t	time;
 
-	time = (int)get_time() - philos->start_time;
-	printf("%d %d %s\n", time, id, str);
+	time = get_time() - philos->start_time;
+	printf("%ld %d %s\n", time, id, str);
 }
 
 void	think(t_philo *philos)
